Fix previous solution
I understand now why they used align center on the .left and .right.
They changed the margin and padding to 0 for the unordered lists
That unaligned the element even though the align-items: center was
applied to the header. Because the unordered list container would be
centered but not the contents in the unordered list. So to align
those contents you would use align center on the left and right
containers instead.