Fix text-align
Most of my code was good, my approach to creating containers
in the HTML was correct. I had slight differences in my HTML code
but ultimately it looked similar. My only issue was aligning the
text to the center and was trying to do that with flex box.
However, it could be done with text-align: center in the body selector
of the css.