Was able to mostly fix css
I was able to get the website to look quite similar to the
desired outcome. However, I couldn't get the footer to the bottom
despite the hint where to said change flex to column.
I was looking at the wrong places and i should've set the flex
on the content to fill up as muchb space as possible to
push the footer down all the way to the bottom.