Complete exercise
I was able to complete the exercise however,
it took me a good while. It took a lot of tinkering
and looking back at the problem however, it was eventually
done. I'm a little worried about when I have to create
my own website or project. Maybe through trial and
error I will improve on my css.